requirements.txt
prefect>=2.20.0
structlog>=24.0.0
boto3>=1.34.0
httpx>=0.27.0
python-dotenv>=1.0.1
pyodbc>=5.1.0
supabase>=2.6.0
psycopg[binary]>=3.2.0

